+++ src/pages/Workshops/special-badge.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import WorkshopsTemplate from './template';
+import { Helmet } from 'react-helmet';
+
+const SpecialBadge = props => {
+  return <WorkshopsTemplate  {...props} />
+}
+
+export default SpecialBadge;
